+++ ui/CHANGELOG.md
@@ -3,6 +3,202 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.16.24](https://github.com/parca-dev/parca/compare/ui-v0.16.23...ui-v0.16.24) (2022-09-27)
+
+
+
+## [0.16.20](https://github.com/parca-dev/parca/compare/ui-v0.16.19...ui-v0.16.20) (2022-09-27)
+
+
+
+## [0.16.19](https://github.com/parca-dev/parca/compare/ui-v0.16.18...ui-v0.16.19) (2022-09-26)
+
+
+
+## [0.16.18](https://github.com/parca-dev/parca/compare/ui-v0.16.17...ui-v0.16.18) (2022-09-26)
+
+
+
+## [0.16.17](https://github.com/parca-dev/parca/compare/ui-v0.16.16...ui-v0.16.17) (2022-09-26)
+
+
+
+## [0.16.16](https://github.com/parca-dev/parca/compare/ui-v0.16.6...ui-v0.16.16) (2022-09-26)
+
+
+### Bug Fixes
+
+* set the granule size option in FrostDB ([#1726](https://github.com/parca-dev/parca/issues/1726)) ([20ba452](https://github.com/parca-dev/parca/commit/20ba452203cfc682f8ce623abe3132904418b694))
+
+
+
+## [0.16.6](https://github.com/parca-dev/parca/compare/ui-v0.16.5...ui-v0.16.6) (2022-09-15)
+
+
+
+## [0.16.5](https://github.com/parca-dev/parca/compare/ui-v0.16.2...ui-v0.16.5) (2022-09-14)
+
+
+
+## [0.16.2](https://github.com/parca-dev/parca/compare/ui-v0.13.4...ui-v0.16.2) (2022-09-14)
+
+
+
+## [0.13.4](https://github.com/parca-dev/parca/compare/ui-v0.13.3...ui-v0.13.4) (2022-06-07)
+
+
+
+## [0.13.3](https://github.com/parca-dev/parca/compare/ui-v0.13.2...ui-v0.13.3) (2022-06-07)
+
+
+
+## [0.13.2](https://github.com/parca-dev/parca/compare/ui-v0.13.1...ui-v0.13.2) (2022-05-31)
+
+
+
+## [0.13.1](https://github.com/parca-dev/parca/compare/ui-v0.13.0...ui-v0.13.1) (2022-05-31)
+
+
+
+# [0.13.0](https://github.com/parca-dev/parca/compare/ui-v0.12.38...ui-v0.13.0) (2022-05-30)
+
+
+### Bug Fixes
+
+* linter errors ([7230db4](https://github.com/parca-dev/parca/commit/7230db46b764d3a11ecb63f2b28a2a666629b5a9))
+
+
+### Features
+
+* useContainerDimensions hook for smooth chart resize ([3fe5670](https://github.com/parca-dev/parca/commit/3fe5670cb94e838d83e5cb10d453ee620c2dc3c1))
+
+
+
+## [0.12.38](https://github.com/parca-dev/parca/compare/ui-v0.12.37...ui-v0.12.38) (2022-05-19)
+
+
+
+## [0.12.37](https://github.com/parca-dev/parca/compare/ui-v0.12.36...ui-v0.12.37) (2022-05-16)
+
+
+
+## [0.12.36](https://github.com/parca-dev/parca/compare/ui-v0.12.35...ui-v0.12.36) (2022-05-06)
+
+
+
+## 0.12.35 (2022-05-05)
+
+
+
+
+
+## [0.16.23](https://github.com/parca-dev/parca/compare/ui-v0.16.20...ui-v0.16.23) (2022-09-27)
+
+**Note:** Version bump only for package ui
+
+
+
+
+
+## [0.16.22](https://github.com/parca-dev/parca/compare/ui-v0.16.20...ui-v0.16.22) (2022-09-27)
+
+**Note:** Version bump only for package ui
+
+
+
+
+
+## [0.16.21](https://github.com/parca-dev/parca/compare/ui-v0.16.20...ui-v0.16.21) (2022-09-27)
+
+**Note:** Version bump only for package ui
+
+
+
+
+
+## [0.16.20](https://github.com/parca-dev/parca/compare/ui-v0.16.19...ui-v0.16.20) (2022-09-27)
+
+
+
+## [0.16.18](https://github.com/parca-dev/parca/compare/ui-v0.16.17...ui-v0.16.18) (2022-09-26)
+
+
+
+## [0.16.17](https://github.com/parca-dev/parca/compare/ui-v0.16.16...ui-v0.16.17) (2022-09-26)
+
+
+
+## [0.16.16](https://github.com/parca-dev/parca/compare/ui-v0.16.15...ui-v0.16.16) (2022-09-26)
+
+
+
+## [0.16.15](https://github.com/parca-dev/parca/compare/ui-v0.16.6...ui-v0.16.15) (2022-09-22)
+
+
+### Bug Fixes
+
+* set the granule size option in FrostDB ([#1726](https://github.com/parca-dev/parca/issues/1726)) ([20ba452](https://github.com/parca-dev/parca/commit/20ba452203cfc682f8ce623abe3132904418b694))
+
+
+
+## 0.16.6 (2022-09-15)
+
+
+
+
+
+## [0.16.19](https://github.com/parca-dev/parca/compare/ui-v0.16.18...ui-v0.16.19) (2022-09-26)
+
+
+
+## [0.16.17](https://github.com/parca-dev/parca/compare/ui-v0.16.16...ui-v0.16.17) (2022-09-26)
+
+
+
+## 0.16.16 (2022-09-26)
+
+**Note:** Version bump only for package ui
+
+
+
+
+
+## [0.16.18](https://github.com/parca-dev/parca/compare/ui-v0.16.17...ui-v0.16.18) (2022-09-26)
+
+
+
+## 0.16.16 (2022-09-26)
+
+**Note:** Version bump only for package ui
+
+
+
+
+
+## [0.16.17](https://github.com/parca-dev/parca/compare/ui-v0.16.16...ui-v0.16.17) (2022-09-26)
+
+**Note:** Version bump only for package ui
+
+
+
+
+
+## [0.16.16](https://github.com/parca-dev/parca/compare/ui-v0.16.15...ui-v0.16.16) (2022-09-26)
+
+**Note:** Version bump only for package ui
+
+
+
+
+
+## [0.16.15](https://github.com/parca-dev/parca/compare/ui-v0.16.14...ui-v0.16.15) (2022-09-22)
+
+**Note:** Version bump only for package ui
+
+
+
+
+
 ## [0.16.14](https://github.com/parca-dev/parca/compare/ui-v0.16.13...ui-v0.16.14) (2022-09-21)
 
 **Note:** Version bump only for package ui